+++ lib/src/extensions/flutter_extensions/build_context.dart
@@ -1,5 +1,0 @@
-import 'package:flutter/material.dart';
-
-extension OtherBuildContextExtension on BuildContext {
-  ScaffoldMessengerState get scaffoldMessenger => ScaffoldMessenger.of(this);
-}