+++ lib/src/extensions/flutter_extensions/flutter_extensions.dart
@@ -1,8 +1,9 @@
 export 'align.dart';
-export 'build_context.dart';
+export 'focus_scope_extensions.dart';
 export 'list_widget.dart';
 export 'media_query_extension.dart';
 export 'navigation.dart';
 export 'padding.dart';
 export 'platform_extension.dart';
+export 'scaffold_messenger_extension.dart';
 export 'theme_extension.dart';