+++ lib/src/src.dart
@@ -1,2 +1,3 @@
+export 'enums/enumd.dart';
 export 'extensions/extensions.dart';
 export 'helper_methods/helper_methods.dart';